Separate ctor tests into separate files by topic
