+++ tests/Moq.Analyzers.Test/ConstructorArgumentsShouldMatchAnalyzerTests.Delegates.cs
@@ -1,0 +1,36 @@
+﻿using Verifier = Moq.Analyzers.Test.Helpers.AnalyzerVerifier<Moq.Analyzers.ConstructorArgumentsShouldMatchAnalyzer>;
+
+namespace Moq.Analyzers.Test;
+public partial class ConstructorArgumentsShouldMatchAnalyzerTests
+{
+    public static IEnumerable<object[]> DelegateTestData()
+    {
+        return new object[][]
+        {
+            ["""new Mock<DelegateWithParam>();"""],
+            ["""new Mock<DelegateWithoutParam>();"""],
+        }.WithNamespaces().WithMoqReferenceAssemblyGroups();
+    }
+
+    [Theory]
+    [MemberData(nameof(DelegateTestData))]
+    public async Task ShouldAnalyzeDelegate(string referenceAssemblyGroup, string @namespace, string mock)
+    {
+        await Verifier.VerifyAnalyzerAsync(
+            $$"""
+              {{@namespace}}
+
+              public delegate void DelegateWithParam(int a);
+              public delegate void DelegateWithoutParam();
+
+              internal class UnitTest
+              {
+                  private void Test()
+                  {
+                      {{mock}}
+                  }
+              }
+              """,
+            referenceAssemblyGroup);
+    }
+}
