+++ tests/Moq.Analyzers.Test/ConstructorArgumentsShouldMatchAnalyzerTests.cs
@@ -2,37 +2,8 @@ using Verifier = Moq.Analyzers.Test.Helpers.AnalyzerVerifier<Moq.Analyzers.Const
 
 namespace Moq.Analyzers.Test;
 
-public class ConstructorArgumentsShouldMatchAnalyzerTests
+public partial class ConstructorArgumentsShouldMatchAnalyzerTests
 {
-    public static IEnumerable<object[]> DelegateTestData()
-    {
-        return new object[][]
-        {
-            ["""new Mock<DelegateWithParam>();"""],
-        }.WithNamespaces().WithMoqReferenceAssemblyGroups();
-    }
-
-    [Theory]
-    [MemberData(nameof(DelegateTestData))]
-    public async Task ShouldAnalyzeDelegate(string referenceAssemblyGroup, string @namespace, string mock)
-    {
-        await Verifier.VerifyAnalyzerAsync(
-            $$"""
-              {{@namespace}}
-
-              public delegate void DelegateWithParam(int a);
-
-              internal class UnitTest
-              {
-                  private void Test()
-                  {
-                      {{mock}}
-                  }
-              }
-              """,
-            referenceAssemblyGroup);
-    }
-
     [System.Diagnostics.CodeAnalysis.SuppressMessage("Design", "MA0051:Method is too long", Justification = "All test cases")]
     public static IEnumerable<object[]> TestData()
     {
@@ -112,158 +83,6 @@ public class ConstructorArgumentsShouldMatchAnalyzerTests
             ["""var repository = new MockRepository(MockBehavior.Default) { DefaultValue = DefaultValue.Empty }; var fooMock = repository.Create<Foo>(false, 42); repository.Verify();"""],
             ["""var repository = new MockRepository(MockBehavior.Default) { DefaultValue = DefaultValue.Empty }; var fooMock = repository.Create<Foo>(MockBehavior.Default, false, 42); repository.Verify();"""],
         }.WithNamespaces().WithMoqReferenceAssemblyGroups();
-    }
-
-    public static IEnumerable<object[]> InterfaceTestData()
-    {
-        IEnumerable<object[]> all = new object[][]
-        {
-            // Regular code (to make sure we bail out early)
-            ["""IFoo foo;"""],
-
-            // Regular
-            ["""new Mock<IFoo>(MockBehavior.Default);"""],
-            ["""new Mock<IFoo>();"""],
-            ["""new Mock<IFoo>{|Moq1001:(MockBehavior.Default, 42)|};"""],
-            ["""new Mock<IFoo>{|Moq1001:(42)|};"""],
-
-            // LINQ
-            ["""Mock.Of<IFoo>();"""],
-
-            // Repository
-            ["""var repository = new MockRepository(MockBehavior.Default) { DefaultValue = DefaultValue.Empty }; var fooMock = repository.Create<IFoo>(MockBehavior.Default); repository.Verify();"""],
-            ["""var repository = new MockRepository(MockBehavior.Default) { DefaultValue = DefaultValue.Empty }; var fooMock = repository.Create<IFoo>(); repository.Verify();"""],
-        }.WithNamespaces().WithMoqReferenceAssemblyGroups();
-
-        IEnumerable<object[]> @new = new object[][]
-        {
-            // LINQ
-            ["""Mock.Of<IFoo>(MockBehavior.Default);"""],   // This is only available in newer versions of
-        }.WithNamespaces().WithNewMoqReferenceAssemblyGroups();
-
-        return all.Union(@new);
-    }
-
-    public static IEnumerable<object[]> ClassWithDefaultCtorTestData()
-    {
-        IEnumerable<object[]> all = new object[][]
-        {
-            ["""new Mock<ClassDefaultCtor>(MockBehavior.Default);"""],
-            ["""new Mock<ClassDefaultCtor>();"""],
-
-            ["""Mock.Of<ClassDefaultCtor>();"""],
-            ["""Mock.Of<ClassDefaultCtor>(m => true);"""],
-
-            ["""var repository = new MockRepository(MockBehavior.Default) { DefaultValue = DefaultValue.Empty }; var fooMock = repository.Create<ClassDefaultCtor>(MockBehavior.Default); repository.Verify();"""],
-            ["""var repository = new MockRepository(MockBehavior.Default) { DefaultValue = DefaultValue.Empty }; var fooMock = repository.Create<ClassDefaultCtor>(); repository.Verify();"""],
-        }.WithNamespaces().WithMoqReferenceAssemblyGroups();
-
-        IEnumerable<object[]> @new = new object[][]
-        {
-            ["""Mock.Of<ClassDefaultCtor>(m => true, MockBehavior.Default);"""],
-            ["""Mock.Of<ClassDefaultCtor>(MockBehavior.Default);"""],
-        }.WithNamespaces().WithNewMoqReferenceAssemblyGroups();
-
-        return all.Union(@new);
-    }
-
-    public static IEnumerable<object[]> ClassWithDefaultParamCtorTestData()
-    {
-        IEnumerable<object[]> all = new object[][]
-        {
-            // Regular
-            ["""new Mock<ClassWithDefaultParamCtor>(MockBehavior.Default);"""],
-            ["""new Mock<ClassWithDefaultParamCtor>();"""],
-            ["""new Mock<ClassWithDefaultParamCtor>(MockBehavior.Default, 21);"""],
-            ["""new Mock<ClassWithDefaultParamCtor>(21);"""],
-
-            // LINQ
-            ["""Mock.Of<ClassWithDefaultParamCtor>();"""],
-
-            // Repository
-            ["""var repository = new MockRepository(MockBehavior.Default) { DefaultValue = DefaultValue.Empty }; var fooMock = repository.Create<ClassWithDefaultParamCtor>(MockBehavior.Default); repository.Verify();"""],
-            ["""var repository = new MockRepository(MockBehavior.Default) { DefaultValue = DefaultValue.Empty }; var fooMock = repository.Create<ClassWithDefaultParamCtor>(); repository.Verify();"""],
-            ["""var repository = new MockRepository(MockBehavior.Default) { DefaultValue = DefaultValue.Empty }; var fooMock = repository.Create<ClassWithDefaultParamCtor>(MockBehavior.Default, 21); repository.Verify();"""],
-            ["""var repository = new MockRepository(MockBehavior.Default) { DefaultValue = DefaultValue.Empty }; var fooMock = repository.Create<ClassWithDefaultParamCtor>(21); repository.Verify();"""],
-        }.WithNamespaces().WithMoqReferenceAssemblyGroups();
-
-        IEnumerable<object[]> @new = new object[][]
-        {
-            // LINQ
-            ["""Mock.Of<ClassWithDefaultParamCtor>(MockBehavior.Default);"""],
-        }.WithNamespaces().WithNewMoqReferenceAssemblyGroups();
-
-        return all.Union(@new);
-    }
-
-    [Theory]
-    [MemberData(nameof(ClassWithDefaultParamCtorTestData))]
-    public async Task ShouldAnalyzeClassWithDefaultParamCtor(string referenceAssemblyGroup, string @namespace, string mock)
-    {
-        await Verifier.VerifyAnalyzerAsync(
-            $$"""
-              {{@namespace}}
-
-              internal class ClassWithDefaultParamCtor
-              {
-                  public ClassWithDefaultParamCtor(int a = 42) { }
-              }
-
-              internal class UnitTest
-              {
-                  private void Test()
-                  {
-                      {{mock}}
-                  }
-              }
-              """,
-            referenceAssemblyGroup);
-    }
-
-    [Theory]
-    [MemberData(nameof(ClassWithDefaultCtorTestData))]
-    public async Task ShouldAnalyzeClassWithDefaultCtor(string referenceAssemblyGroup, string @namespace, string mock)
-    {
-        await Verifier.VerifyAnalyzerAsync(
-            $$"""
-              {{@namespace}}
-
-              public class ClassDefaultCtor
-              {
-              }
-
-              internal class UnitTest
-              {
-                  private void Test()
-                  {
-                      {{mock}}
-                  }
-              }
-              """,
-            referenceAssemblyGroup);
-    }
-
-    [Theory]
-    [MemberData(nameof(InterfaceTestData))]
-    public async Task ShouldAnalyzeInterface(string referenceAssemblyGroup, string @namespace, string mock)
-    {
-        await Verifier.VerifyAnalyzerAsync(
-                $$"""
-                {{@namespace}}
-
-                public interface IFoo
-                {
-                }
-
-                internal class UnitTest
-                {
-                    private void Test()
-                    {
-                        {{mock}}
-                    }
-                }
-                """,
-                referenceAssemblyGroup);
     }
 
     [Theory]
